feat: Complete Project 1 - RISC-V Bootloader & Tiny Kernel (C-Core)
This commit marks the completion of Project 1, implementing a RISC-V
bootloader and a tiny kernel with full C-Core requirements.

Key features and tasks implemented include:

-   **Task 1: Simple Bootloader:** Implemented a basic bootloader
    to print a welcome message.

-   **Task 2: Kernel Initialization:**
    -   Bootloader loads the kernel into memory.
    -   Kernel's `head.S` clears BSS and sets up the stack.
    -   Kernel's `main.c` initializes a jump table, prints "Hello OS!",
        and echoes input.

-   **Task 3 & 4: Application Loading & Tight Packing:**
    -   `createimage` tool enhanced to generate tightly packed
        images, storing applications with their actual sizes.
    -   `task_info_t` extended to include application name, byte
        offset, byte size, and entry point.
    -   Kernel's `loader.c` loads applications by name using the new
        metadata.
    -   `crt0.S` for user applications sets up the C runtime (BSS,
        stack) and returns control to the kernel.

-   **Task 5: Batch Processing (C-Core):**
    -   Implemented a command-line interface (`init/cmd.c`) with
        custom tokenization.
    -   Commands: `help`, `ls`, `exec`, `write_batch`, `exec_batch`.
    -   Batch processing state (mode, current task index, total
        tasks, I/O buffer) is stored persistently in the first sector of
        the image.
    -   `crt0.S` saves application return value (output in `a0`) to
        the persistent I/O buffer and triggers a kernel restart.
    -   Kernel's `main` function handles batch continuation across
        restarts, passing the previous task's output as input to the
        next.
    -   Four number-processing test applications (`app_num1` to
        `app_num4`) created to demonstrate batch I/O passing.

Repository Maintenance:
-   Configured `.gitignore` to properly ignore generated build
    artifacts (`build/`) and editor/IDE cache files (`.cache/`).
-   Removed previously tracked generated files (`build/`,
    `createimage`, `.cache/`) from Git's index to maintain a clean
    repository.

+++ arch/riscv/bios/common.c
@@ -1,5 +1,6 @@
 #include <common.h>
 #include <asm/biosdef.h>
+#include <type.h>
 
 #define BIOS_FUNC_ENTRY 0x50150000
 #define IGNORE 0
@@ -27,7 +28,7 @@ int port_read_ch(void)
     return call_bios((long)BIOS_GETCHAR, IGNORE, IGNORE, IGNORE, IGNORE, IGNORE);
 }
 
-int sd_read(unsigned mem_address, unsigned num_of_blocks, unsigned block_id)
+int sd_read(uintptr_t mem_address, unsigned num_of_blocks, unsigned block_id)
 {
     return (int)call_bios((long)BIOS_SDREAD, (long)mem_address, \
                             (long)num_of_blocks, (long)block_id, IGNORE, IGNORE);
